PLANET-4988 Add C3 probe, setup coverage
* Require c3 probe through composer in the tests dir. The installer puts
c3 in the tests. This is then included by a modified `index.php` in
planet4-docker-compose.
* Split into more job steps, instead of having a single `make` job. That
way we get more detailed timing and job step outputs aren't as long.
* Update lucatume/wp-browser to the latest version. This in turn updates
codeception to 4.1.6, which we need to be able to user the pcov driver
for zero overhead code coverage collection in CI. It also updates
phpunit from 6.5.14 to 8.5.5.
* Some things we use from codeception/codeception where moved to
separate packages, so need to be required separately now.