Adjust production pipeline to run tests on tags
Since we build different dependencies on production, it makes sense to
also run tests on tags and not depend on develop pipeline.

To avoid running the develop pipeline when there is a release scheduled,
added a parameter check to skip ci in this case, when `[release]` prefix
is part of the commit message. We could use built-in `[skip ci]`, but
this would likely skip CI even on tag. And `[release]` is a bit more
clear and explicit.